refactor code so that each component in productOverview widget has its own folder - it makes it easier to navigate the file system
