+++ src/components/productOverview/imageGallery/DefaultView.jsx
