+++ src/components/productOverview/imageGallery/ExpandedView.jsx
