+++ src/components/productOverview/imageGallery/ThumbnailCarousel.jsx
