+++ src/components/productOverview/productInfo/StarReviews.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import React, { useState, useEffect } from 'react';
 import 'whatwg-fetch';
-import { getStars, calculateTotalReviewNumber, calculateAverage } from '../common/helpers.js';
+import { getStars, calculateTotalReviewNumber, calculateAverage } from '../../common/helpers.js';
 
 const StarReviews = ({ currentItem, onReviewLinkClick }) => {
   const [rating, setRating] = useState(0);
