+++ src/components/productOverview/styleSelector/StyleSelector.jsx
