+++ src/components/productOverview/styleSelector/StylesView.jsx
